complete_version/backend/requirements.txt
# StackBridge Backend Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0

# Validation & Settings
pydantic==2.10.4
python-dotenv==1.0.1
